requirements.txt
numpy
pandas
requests
python-dotenv
streamlit==1.32.0
altair<5
